Create image-reflection
Write a function that returns an output image that is twice the height of the original. The
top half of the output image should be identical to the original image. The bottom half,
however, should look like a reflection of the top half. The highest row in the top half
should be “reflected” to be the lowest row in the bottom half. This results in a cool effect.